requirements.txt
flask
flask-compress
flask-session
flask-socketio
flask-sqlalchemy
flask-migrate
flask-pydantic
flask-wtf
flask-login
Flask-WeasyPrint==0.6
passlib
gunicorn
WTForms
WTForms-Alchemy
WTForms-Components
psycopg2-binary
alembic
python-dotenv
pathlib
celery
redis
